Add declaration to evaluate expressions
The new declaration has the form

> #eval <expr>

and evaluates the expression `<expr>`. Can be used while parsing a file
containing declarations to produce some output.
It honours eval tracing options and prints the traces and results after
the file has been fully parsed.

The `#eval` command now also powers the `--repl`, so there is no longer a
special case for expression, but

> 2+2

is syntactic sugar for

> #eval 2+2