Implement Exact Printer for Simala AST
Slightly hacky, but manages to print the exact AST that was written by
the user except for whitespace preceding the first AST node.

We traverse the Simala AST and generate the original LexTokens that have
been used to produce the AST in the first place.

Slightly hackily, we sometimes perform ad-hoc analysis to decide whether
an AST node has a corresponding CST node.